% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationIndex.R
\name{CalcDivIndex}
\alias{CalcDivIndex}
\title{Compute diversity metrics}
\usage{
CalcDivIndex(dataSet, Nvar = "D_cm", Inter = 10, type = "BA")
}
\arguments{
\item{Data.frame}{of population, can gather multiple sites, years and sources}

\item{Type}{string, define wether to compute proportion in terms of frequencies or relative basal area}
}
\value{
A data.frame containing diversity metrics for each site/year/source
}
\description{
This function compute the Gini index for a population
}
