% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{Compute_mingling}
\alias{Compute_mingling}
\title{Compute Species mingling in a plot}
\usage{
Compute_mingling(TabDis, Nk = 4, EdgeCorrection = "NN1")
}
\arguments{
\item{Nk}{int, number of neighboors}

\item{EdgeCorrection}{str, type of edge correction applied}

\item{Tabdis}{object}
}
\value{
data.frame with mingling metrics
}
\description{
This function takes a TabDis object and return mingling metrics
}
