% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{Compute_Size_Diff}
\alias{Compute_Size_Diff}
\title{Compute Size differentiation in a plot}
\usage{
Compute_Size_Diff(TabDis, Nk = 4, EdgeCorrection = "None")
}
\arguments{
\item{TabDis}{object}

\item{Nk}{int, number of neighboors}

\item{EdgeCorrection}{str, type of edge correction applied}
}
\value{
data.frame with size differentiation indexes
}
\description{
This function takes a TabDis object and return mingling metrics
}
