% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{Compute_Winkelmass}
\alias{Compute_Winkelmass}
\title{Compute Uniform angle Index}
\usage{
Compute_Winkelmass(TabDis, Nk = 4)
}
\arguments{
\item{TabDis}{object}

\item{Nk}{int, number of neighboors}

\item{EdgeCorrection}{str, type of edge correction applied}
}
\value{
numeric mean Winkelmass value for the plot
}
\description{
This function takes a TabDis object and return a mean Winkelmass value
}
