% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{ComputeBiodiversity}
\alias{ComputeBiodiversity}
\title{Compute biodiversity Index for a gridded landscape}
\usage{
ComputeBiodiversity(DFGrid)
}
\arguments{
\item{DFgrid, }{Grid object from GridLandcsape function}
}
\value{
diversity indices
}
\description{
This function compute the heterogeneity indices for a gridded landscape
}
