% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{ComputeShScale}
\alias{ComputeShScale}
\title{Compute Shannon Index for a given scale}
\usage{
ComputeShScale(DF, Res = 1)
}
\arguments{
\item{DF, }{data.frame of landscape data}

\item{Res, }{numeric resolution (in km) of the grid}
}
\value{
Diversity indices
}
\description{
This function compute the Shannon index for a landscape and a resolution
}
