% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{createPlot}
\alias{createPlot}
\title{Create a Plot object}
\usage{
createPlot(DF, shape = "quadrat", coord)
}
\arguments{
\item{infile}{Path to the input file}
}
\value{
A matrix of the infile
}
\description{
This function takes a data.frame and return a Plot object. The data.frame
must contain variables: species, D_cm, weight, X, Y
}
