% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{DisToBorder}
\alias{DisToBorder}
\title{Compute distance between trees and border}
\usage{
DisToBorder(Plot)
}
\arguments{
\item{Plot}{object}
}
\value{
A Plot object with DisToBorder variable in the DF
}
\description{
This function takes a Plot object and return a TabDis object
}
