% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResilienceIndex.R
\name{format_samsara}
\alias{format_samsara}
\title{Format Samsara dataset}
\usage{
format_samsara(
  dataRaw,
  ClassInter = 10,
  ClassIni = 7.5,
  Out = "HillNb",
  type = "BA"
)
}
\arguments{
\item{dataRaw, }{data.frame, dataSet containing the simulation output of Samsara}
}
\value{
dataSet, data.table, formatted outuput
}
\description{
This function format the data for further analysis
}
