% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManipulation.R
\name{GatherDataSim}
\alias{GatherDataSim}
\title{Build a csv file gathering observation and simulation for a site}
\usage{
GatherDataSim(evalSite, listsrc = c("data", "landclim", "4c", "salem"))
}
\arguments{
\item{evalSite}{str, name of the site}

\item{listsrc}{str list, name of the subsites}
}
\value{
logical: whether the function could save the output into a csv file
}
\description{
Build a csv file gathering observation and simulation for a site
}
