% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationIndex.R
\name{Gini}
\alias{Gini}
\title{Compute Gini index}
\usage{
Gini(
  Size,
  BA = rep(1, length = length(x)),
  weight = rep(1, length = length(x))
)
}
\arguments{
\item{Size}{numeric vector, size of each treee in the population (in cm)}

\item{BA}{numeric vector, basal area associated with each tree (in m2ha-1)}

\item{Weight, }{numeric vector, weight associated with each tree}
}
\value{
The Gini index for the population
}
\description{
This function compute the Gini index for a population
}
