% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationIndex.R
\name{GiniPop}
\alias{GiniPop}
\title{Compute Gini index}
\usage{
GiniPop(Size, BA, weight = 1, PLOT = FALSE)
}
\arguments{
\item{Size}{numeric vector, size of each treee in the population (in cm)}

\item{BA}{numeric vector, basal area associated with each tree (in m2ha-1)}

\item{PLOT, }{logical, if set to 1 plot the lorenz curve}

\item{Weight, }{numeric vector, weight associated with each tree}
}
\value{
The Gini index for the population
}
\description{
This function compute the Gini index for a population based on the BA proportion
}
