% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{GridLandscape}
\alias{GridLandscape}
\title{Compute a gridded landscape}
\usage{
GridLandscape(DFclass, Res = 1, N = 9)
}
\arguments{
\item{DFclass, }{data.frame of landscape data}

\item{Res, }{numeric resolution (in km) of the grid}

\item{N, }{integer from 1 to 9; here to include different grid configurations}
}
\value{
data.frame of gridded landscape
}
\description{
This function compute the gridded landscape
}
