% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationIndex.R
\name{HillPop}
\alias{HillPop}
\title{Compute Hill Nb}
\usage{
HillPop(Class, weight, OutFormat = "list")
}
\arguments{
\item{Class}{numeric vector, class of each treee in the population (in cm)}

\item{Weight, }{numeric vector, weight associated with each tree, can be the basal area}
}
\value{
The Hill Numbers index for the population
}
\description{
This function compute the Hill numbers index for a population
}
