% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{LandscapeDBHtoClass}
\alias{LandscapeDBHtoClass}
\title{Compute class Landscape}
\usage{
LandscapeDBHtoClass(DF, Nvar = "D_cm", ClassInter = 10, ClassIni = 7.5)
}
\arguments{
\item{DF, }{data.table, the landscape with dbh for each tree}

\item{Nvar, }{string, the variable used to build class, whehter 'species' or 'D_cm'}

\item{ClassInter}{num, Class size (same units as the variable chosen)}

\item{ClassIni}{num, min dbh of the first class}
}
\value{
DFclass, data.table, the equivalent of DF in class form
}
\description{
This function takes a data.table of landscape and return in a class form
}
