% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{plot.DistanceTab}
\alias{plot.DistanceTab}
\title{Plot a TabDis object}
\usage{
plot.DistanceTab(Tdis, Nk = 4)
}
\arguments{
\item{Nk}{int: number of neighboor to exclude/include border plots}

\item{TabDis}{object, discribing ONE plot (one year, one site, one source,...)}
}
\description{
This function makes a plot of a TabDis object
}
