% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{plot.Grid}
\alias{plot.Grid}
\title{Plot a gridded landscape}
\usage{
\method{plot}{Grid}(DFgrid)
}
\arguments{
\item{DFclass, }{data.frame of gridded landscape data}
}
\value{
plot of Basal area
}
\description{
This function compute the gridded landscape
}
