% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResilienceIndex.R
\name{plot.VirtualExperimentEvent}
\alias{plot.VirtualExperimentEvent}
\title{Plot virtual experiment Event perturbation}
\usage{
\method{plot}{VirtualExperimentEvent}(dataSet, Nvar = "BA", RecTime = 20, normalize = "baseline")
}
\arguments{
\item{dataSet, }{formatted dataset (VirtualExperiment object)}

\item{Nvar, }{string, dataSet, variable to be plotted}
}
\value{
plot Variable chosen
}
\description{
This function plot a variable from formatted dataSet for one event perturbation
}
