% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResilienceIndex.R
\name{plot.VirtualExperimentRegime}
\alias{plot.VirtualExperimentRegime}
\title{Plot virtual experiment Regime perturbations}
\usage{
\method{plot}{VirtualExperimentRegime}(dataSet, Nvar = "V_m3", minVar = 0, maxVar = Inf)
}
\arguments{
\item{dataSet, }{formatted dataset (VirtualExperiment object)}

\item{Nvar, }{string, dataSet, variable to be plotted}
}
\value{
plot Variable chosen
}
\description{
This function plot a variable from formatted dataSet for multiple perturbations cases
}
