% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{plotHeterogeneityScale}
\alias{plotHeterogeneityScale}
\title{Compute Heterogeneity metrics for a different scales}
\usage{
plotHeterogeneityScale(DF, Res = c(0.05, 0.1, 1, 2, 5))
}
\arguments{
\item{DF, }{data.frame of distribution landscape data}

\item{Res, }{numeric vector, resolution (in km) of the grid}
}
\value{
Plot
}
\description{
This function compute the Shannon index for a landscape and a resolution
}
