% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataManipulation.R
\name{standVarCalc}
\alias{standVarCalc}
\title{Calculate stand population metrics
This function takes a name of the site and compute for each year the basic population metrics}
\usage{
standVarCalc(evalSite)
}
\arguments{
\item{evalSite}{str, name of the site}
}
\value{
data.frame with calculate yearly aggregated index (N, Dg, BA, BAI, etc) at sp level and stand level (species=="allsp")
}
\description{
Calculate stand population metrics
This function takes a name of the site and compute for each year the basic population metrics
}
