% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{StructuralComplexityIndex}
\alias{StructuralComplexityIndex}
\title{Compute Structural Complexity Index}
\usage{
StructuralComplexityIndex(DF, shape = "quadrat", coord)
}
\arguments{
\item{shape}{str: shape of the plot (quadrat or circular)}

\item{coord}{vect: vector of the coordinates of the plot : (xmin, xmx, ymin, ymax) or (radius)}

\item{data.frame}{DF: dataset}
}
\description{
This function takes a Plot object and return structural complexity index
}
