% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{TriangleSurface}
\alias{TriangleSurface}
\title{Compute triangle surfaces}
\usage{
TriangleSurface(TriDF)
}
\arguments{
\item{deldir}{object}
}
\value{
data.frame with triangle surfaces
}
\description{
This function takes deldir object (plot tesselation) and compute sum of triangle surfaces
}
