% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationIndex.R
\name{CalcDivIndex}
\alias{CalcDivIndex}
\title{Compute diversity metrics}
\usage{
CalcDivIndex(
  dataSet,
  Nvar = "D_cm",
  ClassInter = 10,
  ClassIni = 7.5,
  type = "BA",
  Out = "HillNb"
)
}
\arguments{
\item{dataSet, }{data.frame of population, can gather multiple sites, years and sources}

\item{ClassInter}{num, Class size (same units as the variable chosen)}

\item{ClassIni}{num, min dbh of the first class}

\item{type}{string, define wether to compute proportion in terms of frequencies or relative basal area}

\item{Out}{string, define wether to compute Hillnb or entropy based indices ('HillNb' or 'Entropy'}

\item{N_var}{string, name of the variable used to compute indices}
}
\value{
A data.frame containing diversity metrics for each site/year/source
}
\description{
This function compute the Gini index and Hill numbers for a population
}
