% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{ComputeHeterogeneityScale}
\alias{ComputeHeterogeneityScale}
\title{Compute heterogeneity Index for a distribution landscape and a given resolution}
\usage{
ComputeHeterogeneityScale(DF, Res = 1)
}
\arguments{
\item{DF, }{data.frame of landscape data}

\item{Res, }{numeric resolution (in km) of the grid}
}
\value{
Diversity indices
}
\description{
This function compute the heterogeneity indices for a landscape and a resolution
}
