% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{ComputeSh}
\alias{ComputeSh}
\title{Compute Shannon Index}
\usage{
ComputeSh(DFQuad)
}
\arguments{
\item{DFQuad, }{data.frame of gridded landscape from Quad function}
}
\value{
alpha, beta and gamma diversity based on Shannon Index
}
\description{
This function compute the Shannon index for a gridded landscape
}
