% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LandScapeIndex.R
\name{ComputeSh}
\alias{ComputeSh}
\title{Compute Shannon Index}
\usage{
ComputeSh(DFgrid)
}
\arguments{
\item{DFgrid, }{Grid object from GridLandcsape function}
}
\value{
diversity indices
}
\description{
This function compute the Shannon index for a gridded landscape
}
