% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{createPlot}
\alias{createPlot}
\title{Create a Plot object}
\usage{
createPlot(DF, shape = "quadrat", coord)
}
\arguments{
\item{shape}{str: shape of the plot (quadrat or circular)}

\item{coord}{vect: vector of the coordinates of the plot : (xmin, xmx, ymin, ymax) or (radius)}

\item{data.frame}{DF: dataset}
}
\value{
A Plot object
}
\description{
This function takes a data.frame and return a Plot object. The data.frame
must contain variables: species, D_cm, X, Y
}
