% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResilienceIndex.R
\name{format_Salem}
\alias{format_Salem}
\title{Format Salem dataset}
\usage{
format_Salem(dataRaw)
}
\arguments{
\item{dataRaw, }{data.frame, dataSet containing the simulation output of Salem}
}
\value{
dataSet, data.table, formatted outuput
}
\description{
This function format the data for further analysis
}
