% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResilienceIndex.R
\name{format_salem}
\alias{format_salem}
\title{Format Salem dataset}
\usage{
format_salem(
  dataRaw,
  ClassInter = 10,
  ClassIni = 7.5,
  Out = "HillNb",
  type = "BA"
)
}
\arguments{
\item{dataRaw, }{data.frame, dataSet containing the simulation output of Salem}

\item{ClassInter, }{num, parameter used to compute diversity indices}

\item{ClassIni, }{num, parameter used to compute diversity indices}

\item{Out, }{str, 'HillNb' or 'Entropy'}

\item{type, }{str, parameter used to compute diversity indices}
}
\value{
dataSet, data.table, formatted outuput
}
\description{
This function format the data for further analysis
}
