% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PopulationIndex.R
\name{ReturnDivIndex}
\alias{ReturnDivIndex}
\title{Return Diversity metrics}
\usage{
ReturnDivIndex(evalSite, Nvar = "D_cm", Inter = 10, path = "DATA")
}
\arguments{
\item{evalSite}{string, name of the area (bauges or profound)}

\item{Nvar}{string, name of the diversity axis (D_cm for size, species for species)}

\item{Inter}{int, width of each size class category}

\item{path}{string, path of the data stored}
}
\value{
A data.frame of metrics
}
\description{
This function loads a according to the evalSite value.
}
