% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{StructuralComplexityIndex}
\alias{StructuralComplexityIndex}
\title{Compute Structural Complexity Index}
\usage{
StructuralComplexityIndex(Plot)
}
\arguments{
\item{Plot}{object}
}
\value{
SCI num
}
\description{
This function takes a Plot object and return structural complexity index
}
