% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{TabDist}
\alias{TabDist}
\title{Compute distance between trees in a plot}
\usage{
TabDist(Plot, Nselec = 10)
}
\arguments{
\item{Plot}{object, discribing ONE plot (one year, one site, one source,...)}

\item{Nselec}{int: number of neighboor retained, 10 by default}
}
\value{
A TabDis object
}
\description{
This function takes a Plot object and return a TabDis object
}
