% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotIndex.R
\name{TabDist}
\alias{TabDist}
\title{Compute distance between trees in a plot}
\usage{
TabDist(DF, shape = "quadrat", coord, Nselec = 10, Inter = 10)
}
\arguments{
\item{shape}{str: shape of the plot (quadrat or circular)}

\item{coord}{vect: vector of the coordinates of the plot : (xmin, xmx, ymin, ymax) or (radius)}

\item{Nselec}{int: number of neighboor retained, 10 by default, 10 is the max value}

\item{data.frame}{DF: dataset}

\item{numeric}{Inter: Interval for size classes (10 cm by default)}
}
\value{
A TabDis object
}
\description{
This function takes a Plot object and return a TabDis object
}
